refactor: 인터페이스에서 추상 클래스로 변경
같은 내용이 계속 중복되어 변경합니다.
