+++ src/main/java/domain/participant/Dealer.java
@@ -2,14 +2,11 @@ package domain.participant;
 
 import domain.card.Card;
 import domain.card.CardDeck;
-import java.util.List;
 
-public class Dealer implements Participant {
-
-    private final CardDeck cardDeck;
+public class Dealer extends Participant {
 
     public Dealer() {
-        this.cardDeck = new CardDeck();
+        super(new CardDeck());
     }
 
     public Card getOpenCard() {
@@ -17,27 +14,7 @@ public class Dealer implements Participant {
     }
 
     @Override
-    public void setUpCardDeck(Card first, Card second) {
-        cardDeck.setUpCards(first, second);
-    }
-
-    @Override
-    public boolean canTakeMoreCard(){
+    public boolean canTakeMoreCard() {
         return (calculateScore() <= 16);
-    }
-
-    @Override
-    public void takeMoreCard(Card card) {
-        cardDeck.takeMore(card);
-    }
-
-    @Override
-    public int calculateScore() {
-        return cardDeck.calculateScore();
-    }
-
-    @Override
-    public List<Card> getCards() {
-        return cardDeck.getCards();
     }
 }
