+++ src/main/java/domain/participant/Player.java
@@ -1,20 +1,17 @@
 package domain.participant;
 
-import domain.card.Card;
 import domain.card.CardDeck;
-import java.util.List;
 
-public class Player implements Participant {
+public class Player extends Participant {
 
     private final String name;
-    private final CardDeck cardDeck;
 
     public Player(String name) {
+        super(new CardDeck());
         validateNotBlank(name);
         name = name.trim();
         validateLength(name);
         this.name = name;
-        this.cardDeck = new CardDeck();
     }
 
     private static void validateLength(String name) {
@@ -30,28 +27,8 @@ public class Player implements Participant {
     }
 
     @Override
-    public void setUpCardDeck(Card first, Card second) {
-        cardDeck.setUpCards(first, second);
-    }
-
-    @Override
-    public boolean canTakeMoreCard(){
+    public boolean canTakeMoreCard() {
         return (calculateScore() <= 21);
-    }
-
-    @Override
-    public void takeMoreCard(Card card) {
-        cardDeck.takeMore(card);
-    }
-
-    @Override
-    public int calculateScore() {
-        return cardDeck.calculateScore();
-    }
-
-    @Override
-    public List<Card> getCards() {
-        return cardDeck.getCards();
     }
 
     public String getName() {
